Update ch13-used-cars.R
Changed to fixest and modelsummary where it was meaningful (cumpercent is not available in modelsummary in a reasonable way, therefore it remained in the codes.) I left the stargazer output everywhere. Also there is a duplicate with lm to get the RMSE and other outputs and put into the necessary format for writing out with stargazer. If that is not needed, then one can delete rows: 259-285 and 309-325 (note that these infos are part of etable's output on line 257